wstring path handling for WinAPI CreateProcessW

+++ cppForSwig/BridgeAPI/DBSetup.h
@@ -73,7 +73,7 @@ namespace Armory
          const bool automateDb_;
 
          std::shared_ptr<NetworkPeers::ClientStore> peers_;
-         uint32_t dbPort_ = UINT32_MAX;
+         uint16_t dbPort_ = UINT16_MAX;
          bool hasRun_ = false;
 
          std::string rpcLogin_;
@@ -101,7 +101,7 @@ namespace Armory
 
          bool isDbRunning(void);
          bool isSatoshiRunning(void);
-         uint32_t getDbPort(void) const;
+         uint16_t getDbPort(void) const;
          std::shared_ptr<NetworkPeers::ClientStore> getPeerStore(void) const;
       };
